Extracted zjsunit/output.js
